requirements.txt
macnotesapp==0.8.0
photoscript==0.4.0
py-applescript==1.0.3
